+++ CHANGELOG.rst
@@ -4,6 +4,11 @@
 - Internal: Use @ffi.def_extern instead of @ffi.callback
   `#899 <https://github.com/libgit2/pygit2/issues/899>`_
 
+Deprecations:
+
+- Deprecate ``Repository.create_remote(...)``, use instead
+  ``Repository.remotes.create(...)``
+
 
 1.2.0 (2020-04-05)
 -------------------------